requirements.txt
google_api_python_client==2.65.0
pandas==1.4.3
Pillow==9.3.0
plotly==5.10.0
protobuf==4.21.9
streamlit==1.13.0
streamlit_option_menu==0.3.2
